pyproject.toml
[tool.poetry]
name = "XcelMerger"
version = "0.1.0"
description = "Reto del Bootcamp de Data Science. Aplicación para integrar una única base de datos"
authors = [
   "Sergio Tejedor Moreno", 
   "Carlos García",
   "Afonso Teixeira"
   ]
license = "Apache"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.8.2"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
flask = "^3.0.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = [
   "tests/",
   "__init__.py",
]
mypy_path = [
   ".services/auth/src",
   ".services/frontend/src",
]
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.pytest.ini_options]
testpaths = [
   "services/auth/tests",
   "services/frontend/tests"
]
addopts = [
   "--cov=services/auth/src",
   "--cov=services/frontend/src",
   "--cov-report=term-missing",
   "--cov-report=html"
]
filterwarnings = "ignore::DeprecationWarning"